Code 3 + code 1 fixes
I've implemented code 3 which seems to work fine for everything I've tried throwing at it. There were also a lot of issues with code 1 so I removed the while() loop and changed the layout so that the user has to enter the exact amount of appliances that they said they want to add. I think the changes to code 1 improve readability and also seem to relieve the errors of trying to add too many objects to the inventory array.