Changed menuOptions(), added value check methods
The menuOptions() function is now even MORE readable, and doesn't seem to have any bugs. Take note that all the functions outside the driver that use scanners now must be passed the scanner object as a parameter as this prevents funky stuff from happening (when you have more than one scanner object). The two classes that I added are to make checking if the user entered an integer or double value much easier (we can just call the function instead of writing dumb loops etc). I left code two and the findAppliancesBy() function commented out since they are probably going to have to be altered, but I've gotta go lift some weights up and down so that can be done later.